+++ distribution/lib/Standard/Base/0.0.0-dev/src/Network/URI.enso
@@ -21,8 +21,8 @@ from project.Nothing import all
 
 polyglot java import java.net.URI as Java_URI
 polyglot java import java.net.URISyntaxException
+polyglot java import org.enso.base.enso_cloud.EnsoHeader
 polyglot java import org.enso.base.enso_cloud.EnsoSecretAccessDenied
-polyglot java import org.enso.base.enso_cloud.EnsoSecretHelper.EnsoHeader
 polyglot java import org.enso.base.net.URITransformer
 polyglot java import org.enso.base.net.URIWithSecrets
