+++ std-bits/base/src/main/java/org/enso/base/enso_cloud/EnsoHeader.java
@@ -1,0 +1,12 @@
+package org.enso.base.enso_cloud;
+
+/** An interface for Standard.Base.Network.HTTP.Header.Header can match */
+public interface EnsoHeader {
+  String name();
+
+  EnsoHideableValue hideable_value();
+
+  default HideableValue getValue() {
+    return HideableValue.from(hideable_value());
+  }
+}
