+++ std-bits/base/src/main/resources/META-INF/native-image/org/enso/base/proxy-config.json
@@ -1,9 +1,7 @@
 [
     {
         "interfaces": [
-            "org.enso.base.enso_cloud.AuthenticationProvider$AuthenticationService",
-            "org.enso.base.enso_cloud.EnsoSecretHelper$EnsoHeader",
-            "org.enso.base.enso_cloud.EnsoSecretHelper$EnsoRequestBody"
+            "org.enso.base.enso_cloud.AuthenticationProvider$AuthenticationService"
         ]
     }
 ]
